@@ -1,6 +1,6 @@
 module clash-test
 
-go 1.18
+go 1.19
 
 require (
 	github.com/Dreamacro/clash v0.0.0
