@@ -13,8 +13,8 @@ import (
 	"github.com/Dreamacro/clash/transport/hysteria/pmtud_fix"
 	"github.com/Dreamacro/clash/transport/hysteria/transport"
 	"github.com/lucas-clemente/quic-go"
-	"io/ioutil"
 	"net"
+	"os"
 	"regexp"
 	"strconv"
 	"time"
@@ -52,6 +52,9 @@ func (h *Hysteria) DialContext(ctx context.Context, metadata *C.Metadata, opts .
 		ctx: context.Background(),
 		hyDialer: func() (net.PacketConn, error) {
 			return dialer.ListenPacket(ctx, "udp", "", h.Base.DialOptions(opts...)...)
+		},
+		remoteAddr: func(addr string) (net.Addr, error) {
+			return resolveUDPAddrWithPrefer("udp", addr, h.prefer)
 		},
 	}
 
@@ -135,7 +138,7 @@ func NewHysteria(option HysteriaOption) (*Hysteria, error) {
 	var bs []byte
 	var err error
 	if len(option.CustomCA) > 0 {
-		bs, err = ioutil.ReadFile(option.CustomCA)
+		bs, err = os.ReadFile(option.CustomCA)
 		if err != nil {
 			return nil, fmt.Errorf("hysteria %s load ca error: %w", addr, err)
 		}
@@ -184,11 +187,11 @@ func NewHysteria(option HysteriaOption) (*Hysteria, error) {
 		option.Protocol = DefaultProtocol
 	}
 	if option.ReceiveWindowConn == 0 {
-		quicConfig.InitialStreamReceiveWindow = DefaultStreamReceiveWindow
+		quicConfig.InitialStreamReceiveWindow = DefaultStreamReceiveWindow / 10
 		quicConfig.MaxStreamReceiveWindow = DefaultStreamReceiveWindow
 	}
 	if option.ReceiveWindow == 0 {
-		quicConfig.InitialConnectionReceiveWindow = DefaultConnectionReceiveWindow
+		quicConfig.InitialConnectionReceiveWindow = DefaultConnectionReceiveWindow / 10
 		quicConfig.MaxConnectionReceiveWindow = DefaultConnectionReceiveWindow
 	}
 	if !quicConfig.DisablePathMTUDiscovery && pmtud_fix.DisablePathMTUDiscovery {
@@ -216,12 +219,13 @@ func NewHysteria(option HysteriaOption) (*Hysteria, error) {
 	}
 	return &Hysteria{
 		Base: &Base{
-			name:  option.Name,
-			addr:  addr,
-			tp:    C.Hysteria,
-			udp:   true,
-			iface: option.Interface,
-			rmark: option.RoutingMark,
+			name:   option.Name,
+			addr:   addr,
+			tp:     C.Hysteria,
+			udp:    true,
+			iface:  option.Interface,
+			rmark:  option.RoutingMark,
+			prefer: C.NewDNSPrefer(option.IPVersion),
 		},
 		client: client,
 	}, nil
@@ -287,8 +291,9 @@ func (c *hyPacketConn) WriteTo(p []byte, addr net.Addr) (n int, err error) {
 }
 
 type hyDialerWithContext struct {
-	hyDialer func() (net.PacketConn, error)
-	ctx      context.Context
+	hyDialer   func() (net.PacketConn, error)
+	ctx        context.Context
+	remoteAddr func(host string) (net.Addr, error)
 }
 
 func (h *hyDialerWithContext) ListenPacket() (net.PacketConn, error) {
@@ -297,4 +302,8 @@ func (h *hyDialerWithContext) ListenPacket() (net.PacketConn, error) {
 
 func (h *hyDialerWithContext) Context() context.Context {
 	return h.ctx
+}
+
+func (h *hyDialerWithContext) RemoteAddr(host string) (net.Addr, error) {
+	return h.remoteAddr(host)
 }
