@@ -79,17 +79,23 @@ func (f *Fallback) Unwrap(metadata *C.Metadata) C.Proxy {
 
 func (f *Fallback) findAliveProxy(touch bool) C.Proxy {
 	proxies := f.GetProxies(touch)
-	al := proxies[0]
-	for i := len(proxies) - 1; i > -1; i-- {
-		proxy := proxies[i]
-		if proxy.Name() == f.selected && proxy.Alive() {
-			return proxy
-		}
-		if proxy.Alive() {
-			al = proxy
+	for _, proxy := range proxies {
+		if len(f.selected) == 0 {
+			if proxy.Alive() {
+				return proxy
+			}
+		} else {
+			if proxy.Name() == f.selected {
+				if proxy.Alive() {
+					return proxy
+				} else {
+					f.selected = ""
+				}
+			}
 		}
 	}
-	return al
+
+	return proxies[0]
 }
 
 func (f *Fallback) Set(name string) error {
