@@ -3,6 +3,7 @@ package sniffer
 import (
 	"bytes"
 	"errors"
+	"fmt"
 	C "github.com/Dreamacro/clash/constant"
 	"net"
 	"strings"
@@ -88,13 +89,32 @@ func SniffHTTP(b []byte) (*string, error) {
 			host, _, err := net.SplitHostPort(rawHost)
 			if err != nil {
 				if addrError, ok := err.(*net.AddrError); ok && strings.Contains(addrError.Err, "missing port") {
-					host = rawHost
+					return parseHost(rawHost)
 				} else {
 					return nil, err
 				}
 			}
+
+			if net.ParseIP(host) != nil {
+				return nil, fmt.Errorf("host is ip")
+			}
+
 			return &host, nil
 		}
 	}
 	return nil, ErrNoClue
+}
+
+func parseHost(host string) (*string, error) {
+	if strings.HasPrefix(host, "[") && strings.HasSuffix(host, "]") {
+		if net.ParseIP(host[1:len(host)-1]) != nil {
+			return nil, fmt.Errorf("host is ip")
+		}
+	}
+
+	if net.ParseIP(host) != nil {
+		return nil, fmt.Errorf("host is ip")
+	}
+
+	return &host, nil
 }
